refactor : address model-domain 분리
