+++ blackWhiteDeliveryDriver/src/main/java/com/sparta/blackwhitedeliverydriver/address/service/AddressService.java
@@ -3,11 +3,11 @@ package com.sparta.blackwhitedeliverydriver.address.service;
 import com.sparta.blackwhitedeliverydriver.address.dto.AddressIdResponseDto;
 import com.sparta.blackwhitedeliverydriver.address.dto.AddressRequestDto;
 import com.sparta.blackwhitedeliverydriver.address.dto.AddressResponseDto;
-import com.sparta.blackwhitedeliverydriver.address.entity.Address;
-import com.sparta.blackwhitedeliverydriver.user.entity.User;
+import com.sparta.blackwhitedeliverydriver.model.address.entity.Address;
+import com.sparta.blackwhitedeliverydriver.model.user.entity.User;
 import com.sparta.blackwhitedeliverydriver.user.exception.ExceptionMessage;
-import com.sparta.blackwhitedeliverydriver.address.repository.AddressRepository;
-import com.sparta.blackwhitedeliverydriver.user.repository.UserRepository;
+import com.sparta.blackwhitedeliverydriver.model.address.repository.AddressRepository;
+import com.sparta.blackwhitedeliverydriver.model.user.repository.UserRepository;
 import jakarta.validation.Valid;
 import java.time.LocalDateTime;
 import java.util.UUID;
